Updated
Added a new feature VIPs. VIPs are now need to hire workers to create new tokens. Changed a few other functions and made it so workers have a limited amount rewards they can mint. Instead of having the VIPs in the same file I decided to make a new file for organization purposes, easy editing, and better security. I plan to move the workers to a new file as well making this file exclusively implementation.